Replace equation delim by $$ for correct rendering
This commit replaces the ocurrences of `\begin{equation}` &
`\end{equation}` by `$$` for correct LaTeX rendering.

Context: the markdown parser in blogo only protects the latex blocks that are
specified by `$$` and `$`, it leaves those code blocks untouched, but
for the rest of the code it 'translates' it to HTML.  So for example, if
it has `\begin{equation} A > B \end{equation}`, it will convert it from
markdown to html, outputing: `\begin{equation} A &gt B \end{equation}`,
and then the latex rendering that comes after (in the browser) will
suffer from it for this case with `>` -> `&gt` is not a problem, and the
rendered latex has good appearance, but for more complex latex
combinations the html translation causes the breaks.